Final Outcome (with comments)
THIS IS EDWARD'S EPIC BESTEST BIKES WEBSITE WHICH INCLUDES FEATURES LIKE:

-FILTERS
You can easily filter the price and brand and find the bike of your dreams!

-MESSAGING
Send an email directly to my address with any complaints or queries!

-REVIEWS
You can add even add reviews with images. However since I don't have a way to store reviews in the cloud anywhere the reviews are saved onto yours and truly's local storage on the chosen browser! Ideally use a window that is able to store data onto the local storage thanks!

-BEAUTIFUL BIKES, HELMETS, MERCH AND PARTS!
Have a wee gander at the selection of bikes or buy some of the bestest merch!